add methods and logic to game
add CalculateHandValue method that returns the current value of the hand... add CheckIfValid method that checks to see if the value of the hand is over the upper valid bounds of 21... add CheckForBlackjack method that returns a boolean representing if the hand is exactly 21...

BUG FIX: fix bug in CreateDeck Method in loop condition (change from 0 to 1). Bug allowed for cards of value 0 to be created for all 4 suits.

add some logic in UI flow